% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sGMYC.r
\name{sGMYC}
\alias{sGMYC}
\title{sGMYC: GMYC analysis with subsampling}
\usage{
sGMYC(phy, ntrees = 1, subsamp = 2, nreps = 100, ncores = 1)
}
\arguments{
\item{phy}{Single tree or sample of trees}

\item{ntrees}{Number of trees to analyze from phy (1 to ntrees) (default:1)}

\item{subsamp}{Number of tips to subsample from each delimited species in full analysis (default: 2)}

\item{nreps}{Number of subsampling replicates (default:100)}

\item{ncores}{Number of cores for parallel analysis (when analyzing multiple trees only) (default: 1)}
}
\value{
An object of class “list” with the following elements:
}
\description{
sGMYC: GMYC analysis with subsampling
}
\examples{
data(hypotree)
myres<-sGMYC(hypotree, subsamp=2, nreps=100)
myprobmat<-spec.probmat(myres)
library(bGMYC)
plot.bgmycprobmat(myprobmat,hypotree)
}
\author{
Magalhães, Santos & Camargo
}
