% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.probmat.r
\name{spec.probmat}
\alias{spec.probmat}
\title{spec.probmat}
\usage{
spec.probmat(res)
}
\arguments{
\item{res}{A table containing the reassignment of sequences to the species delimited
with a subsample of sequences when using sGMYC. This table is the element named
'reassignment' in the list returned by sGMYC.}
}
\value{
An object of class “matrix” with the probabilities of conspecificity between all samples
}
\description{
spec.probmat calculates a matrix of probabilities of conspecificity based on the
original spec.promat function of the bGMYC package. When running sGMYC, the
output list contains an table named 'reassignment', which assigns all samples to the
the species delimited with a subsample of sequences.
}
\examples{
data(hypotree)
myres<-sGMYC(hypotree, subsamp=2, nreps=100)
myprobmat<-spec.probmat(myres)
library(bGMYC)
plot.bgmycprobmat(myprobmat,hypotree)
}
\references{
\itemize{
\item Fujisawa T & Barraclough TG. 2013. Delimiting species using single-locus data and the Generalized Mixed Yule Coalescent approach: A revised method and evaluation on simulated data sets. Systematic Biology 62:707–724, doi: 10.1093/sysbio/syt033
\item Pons J, Barraclough TG, Gomez-Zurita J, Cardoso A, Duran DP, Hazell S, Kamoun S, Sumlin WD & Vogler AP. 2006. Sequence-based species delimitation for the DNA taxonomy of undescribed insects. Systematic Biology 55:595-609, doi: 10.1080/10635150600852011.
\item Reid NM & Carstens BC. 2012. Phylogenetic estimation error can decrease the accuracy of species delimitation: a Bayesian implementation of the general mixed Yule-coalescent model. BMC Evolutionary Biology 12:196, doi: 10.1186/1471-2148-12-196.
}
}
\author{
Rafael F. Magalhães, MTT Santos & Arley Camargo
}
